+++ src/Shared/Reflection/TypeResolver/GenericResolver.php
@@ -2,7 +2,6 @@
 
 namespace ArtARTs36\MergeRequestLinter\Shared\Reflection\TypeResolver;
 
-use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\ArrayObjectConverter;
 use ArtARTs36\MergeRequestLinter\Shared\Reflection\Reflector\Type;
 
 final class GenericResolver implements TypeResolver
